requirements.txt
pathlib
click
requests
beautifulsoup4
html5lib
dateparser
python-slugify